feat: test nav from profile
