+++ app/src/main/java/com/github/se/assocify/ui/screens/profile/securityPrivacy/ProfileSecurityPrivacyScreen.kt
@@ -23,11 +23,12 @@ fun ProfileSecurity_PrivacyScreen(navActions: NavigationActions) {
         CenterAlignedTopAppBar(
             title = { Text("Security/Privacy Settings") },
             navigationIcon = {
-              IconButton(onClick = { navActions.back() }, modifier = Modifier.testTag("backButton")) {
-                Icon(
-                    imageVector = Icons.AutoMirrored.Filled.ArrowBack,
-                    contentDescription = "Arrow Back")
-              }
+              IconButton(
+                  onClick = { navActions.back() }, modifier = Modifier.testTag("backButton")) {
+                    Icon(
+                        imageVector = Icons.AutoMirrored.Filled.ArrowBack,
+                        contentDescription = "Arrow Back")
+                  }
             })
       }) {
         Text(
